+++ engine/docs/changelog.md
@@ -1,5 +1,20 @@
 # TBD
 
+# 1.2.0
+### Features
+* The launcher makes a best-effort attempt to pull the latest version of the image being started
+
+### Changes
+* Upgrade to Kurt Core 1.31.1
+* `CreateEnclave` now takes in the version tag of the API container to use, and an emptystring indicates that the engine server should use its own default version
+* The launcher's `Launch` function has been replaced with `LaunchWithCustomVersion` and `LaunchWithDefaultVersion`
+
+### Breaking Changes
+* `CreateEnclaveArgs`'s `api_container_image` property has been replaced with `api_container_image_version_tag`
+    * Users should leave this blank if they want the default API container, and set it to use a custom version of the API container
+* `EngineServerLauncher.Launch` no longer exists
+    * Users should use either `LaunchWithCustomVersion` or `LaunchWithDefaultVersion` depending on their needs
+
 # 1.1.0
 ### Changes
 * Use Node 16.13.0